add support for long long"

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int fib(int n)
 {
-	long res;
+	long long res;
 
 	if ((n == 1) || (n == 2))
 	{
